Updating Distance Calculation
I am breaking a lot of conventions here. This code should probably be in RadialCoordianteCalculations model, the code should probably refactor the storing of the active reference point as a CLLocation object, and I probably should have created my own branch to test this...but ah well doing it this way is easier lol.